+++ include/game/data.h
@@ -1,6 +1,8 @@
 #ifndef DATA_H
 #define DATA_H
 
+#include <map>
+
 struct GameData {
     struct Positions {
 
@@ -12,6 +14,17 @@ struct GameData {
 
     struct Offsets {
         int beatmapStart = 3000;
+        std::map<int, Judgment> judgmentTimings;
+        int maxJudgmentTiming = 200;
+
+        Offsets() {
+            judgmentTimings[-1] = MISS;
+            judgmentTimings[16] = MARVELLOUS;
+            judgmentTimings[43] = PERFECT;
+            judgmentTimings[97] = GREAT;
+            judgmentTimings[133] = GOOD;
+            judgmentTimings[maxJudgmentTiming] = OK;
+        }
     };
 
     ~GameData() {