+++ include/game/menu.h
@@ -8,9 +8,14 @@ class Menu {
 private:
     Game* game;
 
+    std::unordered_map<int, std::unordered_map<std::string, std::filesystem::path>> beatmaps;
+
 public:
     Menu(Game* game);
     ~Menu();
+
+    void loadBeatmaps();
+    void openBeatmap(int id, std::string difficulty);
 };
 
 #endif