+++ CHANGES.md
@@ -1,3 +1,10 @@
+## HOM-1.0.1
+#### ADDED
+- opening beatmap from menu
+- keys in settings
+- key inputs
+- input detection to closest note with judgments
+
 ## HOM-1.0.0
 #### ADDED
 - window